+++ packages/headless/doc-parser/doc-parser.ts
@@ -1,15 +1,26 @@
 import {ApiModel} from '@microsoft/api-extractor-model';
 import {writeFileSync} from 'fs';
-import {resolveController} from './src/controller-resolver';
+import {
+  ControllerConfiguration,
+  resolveController,
+} from './src/controller-resolver';
 
 const apiModel = new ApiModel();
 const apiPackage = apiModel.loadPackage('temp/headless.api.json');
 const entryPoint = apiPackage.entryPoints[0];
 
-const config = {
+const config: ControllerConfiguration = {
   initializer: 'buildPager',
+  samplePaths: {
+    react_class: [
+      'packages/samples/headless-react/src/components/pager/pager.class.tsx',
+    ],
+    react_fn: [
+      'packages/samples/headless-react/src/components/pager/pager.fn.tsx',
+    ],
+  },
 };
 
-const result = resolveController(entryPoint, config.initializer);
+const result = resolveController(entryPoint, config);
 
 writeFileSync('dist/parsed_doc.json', JSON.stringify(result, null, 2));