+++ packages/headless/src/controllers/result-list/headless-result-list.ts
@@ -5,6 +5,7 @@ import {fetchMoreResults} from '../../features/search/search-actions';
 import {registerFieldsToInclude} from '../../features/fields/fields-actions';
 import {Schema, ArrayValue, StringValue} from '@coveo/bueno';
 import {validateOptions} from '../../utils/validate-payload';
+import {buildSearchStatus} from '../search-status/headless-search-status';
 
 const optionsSchema = new Schema<ResultListOptions>({
   fieldsToInclude: new ArrayValue({
@@ -42,6 +43,7 @@ export function buildResultList(
   props?: ResultListProps
 ) {
   const controller = buildController(engine);
+  const searchStatus = buildSearchStatus(engine);
   const {dispatch} = engine;
 
   const options = validateOptions(
@@ -96,10 +98,9 @@ export function buildResultList(
       const state = engine.state;
 
       return {
+        ...searchStatus.state,
         /** The results of the last executed search. */
         results: state.search.results,
-        /** `true` if a search is in progress and `false` otherwise. */
-        isLoading: state.search.isLoading,
       };
     },
     /**